Remove work package activity tab lazy pagination feature flag
https://community.openproject.org/work_packages/71141
