--- app/components/work_packages/activities_tab/index_component.html.erb
@@ -1,53 +1,0 @@
-<%=
-  if deferred
-    render(
-      WorkPackages::ActivitiesTab::Journals::IndexComponent.new(work_package:, filter:, deferred:)
-    )
-  else
-    component_wrapper(tag: "turbo-frame") do
-      flex_layout(classes: "work-packages-activities-tab-index-component") do |activities_tab_wrapper_container|
-        activities_tab_wrapper_container.with_row(classes: "work-packages-activities-tab-index-component--errors") do
-          render(
-            WorkPackages::ActivitiesTab::ErrorStreamComponent.new
-          )
-        end
-
-        activities_tab_wrapper_container.with_row(id: index_content_wrapper_key, data: wrapper_data_attributes) do
-          flex_layout do |activities_tab_container|
-            activities_tab_container.with_row(mb: 2) do
-              render(
-                WorkPackages::ActivitiesTab::Journals::FilterAndSortingComponent.new(
-                  work_package:,
-                  filter:
-                )
-              )
-            end
-
-            activities_tab_container.with_row(flex_layout: true, classes: "work-packages-activities-tab-index-component--content-container", mt: 3) do |journals_wrapper_container|
-              journals_wrapper_container.with_row(
-                classes: "work-packages-activities-tab-index-component--journals-container work-packages-activities-tab-index-component--journals-container_with-initial-input-compensation",
-                data: { "work-packages--activities-tab--index-target": "journalsContainer" }
-              ) do
-                render(list_journals_component)
-              end
-
-              if adding_comment_allowed?
-                journals_wrapper_container.with_row(
-                  id: add_comment_wrapper_key,
-                  classes: "work-packages-activities-tab-index-component--input-container work-packages-activities-tab-index-component--input-container_sort-#{journal_sorting}",
-                  p: 3,
-                  bg: :subtle,
-                  data: add_comment_wrapper_data_attributes
-                ) do
-                  render(
-                    WorkPackages::ActivitiesTab::Journals::NewComponent.new(work_package:, filter:, last_server_timestamp:)
-                  )
-                end
-              end
-            end
-          end
-        end
-      end
-    end
-  end
-%>
