--- app/components/work_packages/activities_tab/journals/index_component.html.erb
@@ -1,76 +1,0 @@
-<%=
-  if deferred
-    helpers.turbo_frame_tag("work-package-activities-tab-content-older-journals") do
-      flex_layout do |older_journals_container|
-        older_journals.each do |record|
-          older_journals_container.with_row do
-            if record.is_a?(Changeset)
-              render(WorkPackages::ActivitiesTab::Journals::RevisionComponent.new(changeset: record, filter:))
-            else
-              render(
-                WorkPackages::ActivitiesTab::Journals::ItemComponent.new(
-                  journal: record, filter:,
-                  grouped_emoji_reactions: wp_journals_grouped_emoji_reactions[record.id]
-                )
-              )
-            end
-          end
-        end
-      end
-    end
-  else
-    component_wrapper(class: "work-packages-activities-tab-journals-index-component") do
-      flex_layout(data: { test_selector: "op-wp-journals-#{filter}-#{journal_sorting}" }) do |journals_index_wrapper_container|
-        journals_index_wrapper_container.with_row(
-          classes: "work-packages-activities-tab-journals-index-component--journals-inner-container",
-          mb: inner_container_margin_bottom
-        ) do
-          flex_layout(
-            id: insert_target_modifier_id,
-            data: { test_selector: "op-wp-journals-container" }
-          ) do |journals_index_container|
-            if empty_state?
-              journals_index_container.with_row(mt: 2, mb: 3) do
-                render(
-                  WorkPackages::ActivitiesTab::Journals::EmptyComponent.new
-                )
-              end
-            end
-
-            if !journal_sorting.desc? && journals.count > MAX_RECENT_JOURNALS
-              journals_index_container.with_row do
-                helpers.turbo_frame_tag("work-package-activities-tab-content-older-journals", src: work_package_activities_path(work_package, filter:, deferred: true))
-              end
-            end
-
-            recent_journals.each do |record|
-              journals_index_container.with_row do
-                if record.is_a?(Changeset)
-                  render(WorkPackages::ActivitiesTab::Journals::RevisionComponent.new(changeset: record, filter:))
-                else
-                  render(
-                    WorkPackages::ActivitiesTab::Journals::ItemComponent.new(
-                      journal: record, filter:,
-                      grouped_emoji_reactions: wp_journals_grouped_emoji_reactions[record.id]
-                    )
-                  )
-                end
-              end
-            end
-
-            if journal_sorting.desc? && journals.count > MAX_RECENT_JOURNALS
-              journals_index_container.with_row do
-                helpers.turbo_frame_tag("work-package-activities-tab-content-older-journals", src: work_package_activities_path(work_package, filter:, deferred: true))
-              end
-            end
-          end
-        end
-
-        unless empty_state?
-          journals_index_wrapper_container
-            .with_row(classes: "work-packages-activities-tab-journals-index-component--stem-connection")
-        end
-      end
-    end
-  end
-%>
