+++ app/controllers/work_packages/activities_tab_controller.rb
@@ -39,29 +39,19 @@ class WorkPackages::ActivitiesTabController < ApplicationController
   before_action :find_journal, only: %i[emoji_actions item_actions edit cancel_edit update toggle_reaction]
   before_action :set_filter
   before_action :authorize
-  before_action :initialize_pagination, only: %i[page_streams]
+  before_action :initialize_pagination, only: %i[index page_streams]
 
   def index
-    index_component =
-      if OpenProject::FeatureDecisions.wp_activity_tab_lazy_pagination_active?
-        initialize_pagination
-        WorkPackages::ActivitiesTab::LazyIndexComponent.new(
-          work_package: @work_package,
-          journals: @paginated_journals,
-          paginator: @paginator,
-          filter: @filter,
-          last_server_timestamp: get_current_server_timestamp
-        )
-      else
-        WorkPackages::ActivitiesTab::IndexComponent.new(
-          work_package: @work_package,
-          filter: @filter,
-          last_server_timestamp: get_current_server_timestamp,
-          deferred: ActiveRecord::Type::Boolean.new.cast(params[:deferred])
-        )
-      end
-
-    render(index_component, layout: false)
+    render(
+      WorkPackages::ActivitiesTab::LazyIndexComponent.new(
+        work_package: @work_package,
+        journals: @paginated_journals,
+        paginator: @paginator,
+        filter: @filter,
+        last_server_timestamp: get_current_server_timestamp
+      ),
+      layout: false
+    )
   end
 
   def page_streams
@@ -323,37 +313,28 @@ class WorkPackages::ActivitiesTabController < ApplicationController
   end
 
   def replace_whole_tab
-    component =
-      if OpenProject::FeatureDecisions.wp_activity_tab_lazy_pagination_active?
-        initialize_pagination # re-initialize pagination to pick up changes to sorting/filtering
-        WorkPackages::ActivitiesTab::LazyIndexComponent.new(
-          work_package: @work_package,
-          journals: @paginated_journals,
-          paginator: @paginator,
-          filter: @filter,
-          last_server_timestamp: get_current_server_timestamp
-        )
-      else
-        WorkPackages::ActivitiesTab::IndexComponent.new(
-          work_package: @work_package,
-          filter: @filter,
-          last_server_timestamp: get_current_server_timestamp
-        )
-      end
-    replace_via_turbo_stream(component:)
+    initialize_pagination # re-initialize pagination to pick up changes to sorting/filtering
+    replace_via_turbo_stream(
+      component: WorkPackages::ActivitiesTab::LazyIndexComponent.new(
+        work_package: @work_package,
+        journals: @paginated_journals,
+        paginator: @paginator,
+        filter: @filter,
+        last_server_timestamp: get_current_server_timestamp
+      )
+    )
   end
 
   def update_index_component
-    component =
-      if OpenProject::FeatureDecisions.wp_activity_tab_lazy_pagination_active?
-        initialize_pagination # re-initialize pagination to pick up changes to sorting/filtering
-        WorkPackages::ActivitiesTab::Journals::LazyIndexComponent
-        .new(work_package: @work_package, journals: @paginated_journals, paginator: @paginator, filter: @filter)
-      else
-        WorkPackages::ActivitiesTab::Journals::IndexComponent
-          .new(work_package: @work_package, filter: @filter)
-      end
-    update_via_turbo_stream(component:)
+    initialize_pagination # re-initialize pagination to pick up changes to sorting/filtering
+    update_via_turbo_stream(
+      component: WorkPackages::ActivitiesTab::Journals::LazyIndexComponent.new(
+        work_package: @work_package,
+        journals: @paginated_journals,
+        paginator: @paginator,
+        filter: @filter
+      )
+    )
   end
 
   def create_journal_service_call
@@ -421,56 +402,27 @@ class WorkPackages::ActivitiesTabController < ApplicationController
   end
 
   def rerender_journals_with_updated_notification(journals, last_update_timestamp, grouped_emoji_reactions, editing_journals)
-    # Case: the user marked the journal as read somewhere else and expects the bubble to disappear
-    #
-    # below code stopped working with the introduction of the sequence_version query
-    # I believe it is due to the fact that the notification join does not work well with the sequence_version query
-    # see below comments from my debugging session
-    # journals
-    #   .joins(:notifications)
-    #   .where("notifications.updated_at > ?", last_update_timestamp)
-    #   .find_each do |journal|
-    #   # DEBUGGING:
-    #   # the journal id is actually 85 but below data is logged:
-    #   # # journal id 14 (?!)
-    #   # # journal sequence_version 22 (correct!)
-    #   # the update stream has a wrong target then!
-    #   # target="work-packages-activities-tab-journals-item-component-14"
-    #   # instead of
-    #   # target="work-packages-activities-tab-journals-item-component-85"
-    #   update_item_show_component(journal:, grouped_emoji_reactions: grouped_emoji_reactions.fetch(journal.id, {}))
-    # end
-    #
-    # alternative approach in order to bypass the notification join issue in relation with the sequence_version query
     Notification
       .where(journal_id: journals.pluck(:id))
       .where(recipient_id: User.current.id)
       .where("notifications.updated_at > ?", last_update_timestamp)
       .find_each do |notification|
-      next if editing_journals.include?(notification.journal_id)
+        next if editing_journals.include?(notification.journal_id)
 
-      update_item_show_component(
-        journal: journals.find(notification.journal_id), # take the journal from the journals querried with sequence_version!
-        grouped_emoji_reactions: grouped_emoji_reactions.fetch(notification.journal_id, {})
-      )
-    end
+        update_item_show_component(
+          journal: journals.find(notification.journal_id), # take the journal from the journals querried with sequence_version!
+          grouped_emoji_reactions: grouped_emoji_reactions.fetch(notification.journal_id, {})
+        )
+      end
   end
 
   def insert_latest_journals_via_turbo_stream(journals, last_update_timestamp, emoji_reactions)
-    target_component =
-      if OpenProject::FeatureDecisions.wp_activity_tab_lazy_pagination_active?
-        WorkPackages::ActivitiesTab::Journals::LazyIndexComponent.new(
-          work_package: @work_package,
-          journals: Journal.none, # we do not need to pass any journals here since we just want the component key
-          paginator: nil,
-          filter: @filter
-        )
-      else
-        WorkPackages::ActivitiesTab::Journals::IndexComponent.new(
-          work_package: @work_package,
-          filter: @filter
-        )
-      end
+    target_component = WorkPackages::ActivitiesTab::Journals::LazyIndexComponent.new(
+      work_package: @work_package,
+      journals: Journal.none, # we do not need to pass any journals here since we just want the component key
+      paginator: nil,
+      filter: @filter
+    )
 
     journals.where("created_at > ?", last_update_timestamp).find_each do |journal|
       insert_via_turbo_stream(
